Improve product grid display

+++ tkpdshop-common/src/main/java/com/tokopedia/shop/common/data/source/cloud/model/ShopInfoQualityWidth.java
@@ -8,7 +8,7 @@ public class ShopInfoQualityWidth {
 
     @SerializedName("five_star_rank")
     @Expose
-    private long fiveStarRank;
+    private double fiveStarRank;
     @SerializedName("four_star_rank")
     @Expose
     private double fourStarRank;
@@ -22,11 +22,11 @@ public class ShopInfoQualityWidth {
     @Expose
     private double twoStarRank;
 
-    public long getFiveStarRank() {
+    public double getFiveStarRank() {
         return fiveStarRank;
     }
 
-    public void setFiveStarRank(long fiveStarRank) {
+    public void setFiveStarRank(double fiveStarRank) {
         this.fiveStarRank = fiveStarRank;
     }
